requirements.txt
pyserial==3.5
pygame==2.1.2
tweepy==4.13.0
